+++ dart/lib/trinsic.dart
@@ -3,6 +3,11 @@
 /// More dartdocs go here.
 library trinsic_dart;
 
-export 'src/dart_base.dart';
+export 'src/account_service.dart';
+export 'src/credential_service.dart';
+export 'src/provider_service.dart';
+export 'src/template_service.dart';
+export 'src/trustregistry_service.dart';
+export 'src/wallet_service.dart';
 
 // TODO: Export any libraries intended for clients of this package.
